Amilca's CA
Emmanuel, I did 781 lines of code, please have a look on it code and modify if it is necessary. As you know, the way as lecture write the brief for CAs is confused and is hard to understandable. If I am right I did 50% the code I hope you can complete the task for now on. Also, I have been doing it in bad helth condition.

We will keep in touch.
Best regards.

Co-Authored-By: misleinecosta <misleinecosta@users.noreply.github.com>